requirements.txt
Flask==2.0.1
pymongo==4.1.1
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-Session==0.4.0 